Extract TaggedOffsetProducerActor to a new class
